+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Remove scalar field generator [#100]
+- Remove canonical and canonical_derive dependency [#108]
 
 ## [0.11.3] - 2023-05-17
 
@@ -185,6 +186,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename `S` to `TWO_ADACITY` and export it
 
 <!-- Issues -->
+[#108]: (https://github.com/dusk-network/bls12_381/issues/108)
 [#100]: (https://github.com/dusk-network/bls12_381/issues/100)
 [#93]: (https://github.com/dusk-network/bls12_381/issues/93)
 [#75]: (https://github.com/dusk-network/bls12_381/issues/75)